+++ src/Screen/Fields/Upload.php
@@ -136,7 +136,7 @@ class Upload extends Field
             throw_if(
                 $maxFileSize > $serverMaxFileSize,
                 \RuntimeException::class,
-                'Cannot set the desired maximum file sizw. This contradicts the settings specified in .ini');
+                'Cannot set the desired maximum file size. This contradicts the settings specified in .ini');
         });
 
         // set load relation attachment