+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 6.9.0 - 2020-01-27
+
+### Added
+
+- Method `maxFileSize` for `Cropper` and `Picture` fields
+- Method `hiddenEmpty` for `ItemMenu`
+
 ## 6.8.6 - 2020-01-25
 
 ### Fixed